+++ src/Components/Inspirations/imageApiCalls.js
@@ -1,4 +1,4 @@
-import './Components/Inspirations/Inspirations.scss'
+import './Inspirations.scss'
 
 export const getImages = () => {
   return fetch('https://picsum.photos/v2/list')
